+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## Unreleased
+## 0.3.11 - 2026-06-12
+
+- Updated pdfkit to 0.19.1 (fixed RGB JPEG embedded as DeviceGray, bug introduced in 0.19.0)
+
+## 0.3.10 - 2026-06-07
+
+- Updated pdfkit to 0.19.0
+
+## 0.3.9 - 2026-05-23
+
+- Fixed table rowSpan cell height in last column
+
+## 0.3.8 - 2026-05-06
 
 - Added server-side method `setLocalAccessPolicy()` for defining a custom access policy for local file
 
@@ -11,8 +23,10 @@
 		return path.startsWith("fonts/");
 	});
 	```
+- Improved URL Access Policy (`setUrlAccessPolicy` method); URLs are now validated even before each redirection in Node.js and after final redirection in browser (browsers do not support validation before redirection)
 - Fixed extra blank page when using headerRows, dontBreakRows and cell pageBreak together
 - Fixed rendering of an invalid color name - previously it used the last valid color; now it defaults to black
+- Fixed dontBreakRows rowSpan ending offset across pages
 - Added dynamic `pageMargins`
 
 ## 0.3.7 - 2026-03-17